% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_emax.R
\name{print.boot_emax}
\alias{print.boot_emax}
\title{Print a \code{boot_emax} object}
\usage{
\method{print}{boot_emax}(x, ...)
}
\arguments{
\item{x}{A \code{boot_emax} object.}

\item{...}{Currently unused; included for S3 compatibility.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{boot_emax} object
}
\seealso{
\code{\link[=bootstrap_emax]{bootstrap_emax()}}, \code{\link[=summary.boot_emax]{summary.boot_emax()}}
}
